abstract artist and label common functionalities
