+++ Domain/Catalog/Exceptions/Common/ArtistLabelException.cs
@@ -1,0 +1,11 @@
+﻿using SimpleMusicStore.Domain.Common;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SimpleMusicStore.Domain.Catalog.Exceptions.Common
+{
+    public class ArtistLabelException : BaseDomainException
+    {
+    }
+}
