+++ Domain/Catalog/Exceptions/InvalidLabelException.cs
@@ -1,8 +1,9 @@
 ﻿namespace SimpleMusicStore.Domain.Catalog.Exceptions
 {
     using Common;
+    using SimpleMusicStore.Domain.Common;
 
-    public class InvalidLabelException : BaseDomainException
+    public class InvalidLabelException : ArtistLabelException
     {
         public InvalidLabelException()
         {
