+++ Domain/Catalog/Exceptions/InvalidPriceException.cs
@@ -1,6 +1,7 @@
 ﻿namespace SimpleMusicStore.Domain.Catalog.Exceptions
 {
     using Common;
+    using SimpleMusicStore.Domain.Common;
 
     public class InvalidPriceException : BaseDomainException
     {
